requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Pillow==10.1.0
email-validator==2.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
werkzeug==3.0.1
bleach==6.1.0
pytest==8.2.0
pytest-flask==1.3.0
